# perf: Use a global tokio runtime