--- a/common/src/main/java/org/apache/comet/parquet/Native.java
+++ b/common/src/main/java/org/apache/comet/parquet/Native.java
@@ -257,9 +257,7 @@ public final class Native extends NativeBase {
       byte[] requiredSchema,
       byte[] dataSchema,
       String sessionTimezone,
-      int batchSize,
-      int workerThreads,
-      int blockingThreads);
+      int batchSize);
 
   // arrow native version of read batch
   /**