--- a/common/src/main/java/org/apache/comet/parquet/NativeBatchReader.java
+++ b/common/src/main/java/org/apache/comet/parquet/NativeBatchReader.java
@@ -357,16 +357,6 @@ public class NativeBatchReader extends RecordReader<Void, ColumnarBatch> impleme
         conf.getInt(
             CometConf.COMET_BATCH_SIZE().key(),
             (Integer) CometConf.COMET_BATCH_SIZE().defaultValue().get());
-    int workerThreads =
-        conf.getInt(
-            CometConf.COMET_WORKER_THREADS().key(),
-            (Integer) CometConf.COMET_WORKER_THREADS().defaultValue().get());
-    ;
-    int blockingThreads =
-        conf.getInt(
-            CometConf.COMET_BLOCKING_THREADS().key(),
-            (Integer) CometConf.COMET_BLOCKING_THREADS().defaultValue().get());
-    ;
     this.handle =
         Native.initRecordBatchReader(
             filePath,
@@ -377,9 +367,7 @@ public class NativeBatchReader extends RecordReader<Void, ColumnarBatch> impleme
             serializedRequestedArrowSchema,
             serializedDataArrowSchema,
             timeZoneId,
-            batchSize,
-            workerThreads,
-            blockingThreads);
+            batchSize);
     isInitialized = true;
   }