## Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

Closes https://github.com/apache/datafusion-comet/issues/1590

Helps with https://github.com/apache/datafusion-comet/issues/1523

## Rationale for this change

<!--
 Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
 Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.  
-->

If I configure Spark executor with 8GB + 1GB offHeap and using default settings for tokio threads, q4 hangs in main branch, but completes with these changes.

## What changes are included in this PR?

<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

- Allocate one global tokio runtime per process (executor) rather than create one runtime per query
- Use tokio defaults for number of threads
- Allow tokio thread counts to be configured with env vars

Based on an executor configured for 8 concurrent tasks and 1 core per task, the defaults for the overall process now change as follows:

| Config | Before | After |
|-|-|-|
| worker threads | 32 | 8 |
| max blocking threads | 80 | 512 |

## How are these changes tested?

<!--
We typically require tests for all PRs in order to:
1. Prevent the code from being accidentally broken by subsequent changes
2. Serve as another way to document the expected behavior of the code

If tests are not included in your PR, please explain why (for example, are they covered by existing tests)?
-->

Manual testing. I do not see any change in overall TPC-H performance, but I can now get q4 to complete with less memory than before.

Note that we cannot close #1523 until https://github.com/apache/datafusion/issues/15323 is resolved.